Add optional health/telemetry capability
Adding telemetry in order to add health checks and aid in debugging.
This is disabled by default and the structures returned are alpha
quality so that we can better understand the practical usage of them.
